+++ org.sonarlint.eclipse.cdt/META-INF/MANIFEST.MF
@@ -12,6 +12,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.cdt.core;resolution:=optional,
  org.eclipse.cdt.ui;resolution:=optional,
  org.sonarlint.eclipse.core,
+ org.sonarlint.eclipse.ui,
  org.eclipse.core.filesystem,
  org.eclipse.jdt.annotation;resolution:=optional,
  org.eclipse.text,
