+++ org.sonarlint.eclipse.cdt/plugin.xml
@@ -13,7 +13,7 @@
         </provider>
      </extension>
      <extension
-         point="org.sonarlint.eclipse.core.syntaxHighlightingProvider">
+         point="org.sonarlint.eclipse.ui.syntaxHighlightingProvider">
       <enhancer
             class="org.sonarlint.eclipse.cdt.internal.CProjectConfiguratorExtension">
       </enhancer>
