+++ org.sonarlint.eclipse.core/plugin.xml
@@ -8,9 +8,6 @@
    <extension-point id="analysisConfigurator" name="SonarLint Analysis Configurator" schema="schema/analysisConfigurator.exsd"/>
    <extension-point id="languageProvider" name="SonarLint File Language Provider" schema="schema/languageProvider.exsd"/>
    <extension-point id="typeProvider" name="SonarLint File Type Provider" schema="schema/typeProvider.exsd"/>
-   <extension-point id="syntaxHighlightingProvider"
-   					name="SonarLint Syntax Highlighting Provider"
-   					schema="schema/syntaxHighlightingProvider.exsd"/>
    <extension-point id="projectHierarchyProvider"
    					name="SonarLint Project Hierarchy Provider"
    					schema="schema/projectHierarchyProvider.exsd" />
