+++ org.sonarlint.eclipse.core/src/org/sonarlint/eclipse/core/internal/extension/SonarLintExtensionTracker.java
@@ -25,11 +25,10 @@ import org.sonarlint.eclipse.core.analysis.IAnalysisConfigurator;
 import org.sonarlint.eclipse.core.analysis.IFileLanguageProvider;
 import org.sonarlint.eclipse.core.analysis.IFileTypeProvider;
 import org.sonarlint.eclipse.core.configurator.ProjectConfigurator;
-import org.sonarlint.eclipse.core.resource.ISonarLintProjectHierarchyProvider;
 import org.sonarlint.eclipse.core.resource.ISonarLintFileAdapterParticipant;
 import org.sonarlint.eclipse.core.resource.ISonarLintProjectAdapterParticipant;
+import org.sonarlint.eclipse.core.resource.ISonarLintProjectHierarchyProvider;
 import org.sonarlint.eclipse.core.resource.ISonarLintProjectsProvider;
-import org.sonarlint.eclipse.core.rule.ISyntaxHighlightingProvider;
 
 public class SonarLintExtensionTracker extends AbstractSonarLintExtensionTracker {
 
@@ -43,13 +42,11 @@ public class SonarLintExtensionTracker extends AbstractSonarLintExtensionTracker
     "org.sonarlint.eclipse.core.projectAdapterParticipant"); //$NON-NLS-1$
   private final SonarLintEP<IFileLanguageProvider> languageEp = new SonarLintEP<>("org.sonarlint.eclipse.core.languageProvider"); //$NON-NLS-1$
   private final SonarLintEP<IFileTypeProvider> typeEp = new SonarLintEP<>("org.sonarlint.eclipse.core.typeProvider"); //$NON-NLS-1$
-  private final SonarLintEP<ISyntaxHighlightingProvider> syntaxHighlightingProviderEP = new SonarLintEP<>(
-    "org.sonarlint.eclipse.core.syntaxHighlightingProvider"); //$NON-NLS-1$
   private final SonarLintEP<ISonarLintProjectHierarchyProvider> projectHierarchyProviderEP = new SonarLintEP<>(
     "org.sonarlint.eclipse.core.projectHierarchyProvider"); //$NON-NLS-1$
 
   private final Collection<SonarLintEP<?>> allEps = List.of(configuratorEp, analysisEp, projectsProviderEp, fileAdapterParticipantEp, projectAdapterParticipantEp,
-    languageEp, typeEp, syntaxHighlightingProviderEP, projectHierarchyProviderEP);
+    languageEp, typeEp, projectHierarchyProviderEP);
 
   private SonarLintExtensionTracker() {
     init(allEps);
@@ -94,10 +91,6 @@ public class SonarLintExtensionTracker extends AbstractSonarLintExtensionTracker
 
   public Collection<IFileTypeProvider> getTypeProviders() {
     return typeEp.getInstances();
-  }
-
-  public Collection<ISyntaxHighlightingProvider> getSyntaxHighlightingProvider() {
-    return syntaxHighlightingProviderEP.getInstances();
   }
 
   public Collection<ISonarLintProjectHierarchyProvider> getProjectHierarchyProviders() {
