+++ org.sonarlint.eclipse.pydev/plugin.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <plugin>
    <extension
-         point="org.sonarlint.eclipse.core.syntaxHighlightingProvider">
+         point="org.sonarlint.eclipse.ui.syntaxHighlightingProvider">
       <enhancer
             class="org.sonarlint.eclipse.pydev.internal.PythonProjectConfiguratorExtension">
       </enhancer>
